Add branching cords
When a controller goes near an existing cord a green dot called a branch
handle appears at the point on the cord which is closest to the
controller.

The user can grab this handle in order to get a cord that diverges from
the original cord at the point where the handle is.

There is a notion of data flow along the cords. Data flows from output
jacks to input jacks. Data will never flow backwards, so if an input is
connected upstream from an output then the data from the output will not
reach that input.