Smooth out branch handle movement along cord
Movement is still somewhat jittery due to the imprecise calculation of the
closest point on the cord. This could be improved by calculating the
neighboring cord point which is closest to the controller and estimating the true closest point
by interpolating between the closest cord point and the closest neighbor.