Add circular buffer based deque implementation
The new note selector will need a data structure that allows for fast
push/pop at both ends as well as access to the middle element. While a
doubly-linked list would satisfy this requirement, a circular buffer
based deque is more suited to the problem since the size will not need
to change, so we can achieve zero allocation.