Add clipping planes, note display to new NoteSelector
All of the key materials have a clipping plane set by the NoteSelector
so that no part of any key is rendered outside the selector's frame.
This makes it look like the keys are being consumed by the frame as they
move into it.

A note display has been added to the frame so that the user can see the
note name and octave of the selected note.

Finally, out of range keys have been given a new material in an attempt
to make it less confusing for users that reach the limit of the MIDI
note range.

Note: since the clipping plane is a global setting, there may only be
one instance of the note selector at a time. It would make sense for
NoteOrbs to try and lock the note selector when they are grabbed and release
it when they are ungrabbed.