Refactor connector components, close #23
PlugAttach is the new abstract base class for all Plug attachment
mechanisms. SocketPlugAttach is the first concrete implementation of
this class and functions identically to the old Plug/Jack system.

PlugReceptacle is a generic physical data endpoint.

A PlugAttach detects nearby Plugs and attempts to get a lock on one.
Getting this lock means that the PlugAttach has control over the Plug.
It controls how it can move as well as when a data connection to the
PlugReceptacle is made. All of this is left to concrete implementations
of PlugAttach.

In the example of SocketPlugAttach, the locked Plug is snapped to the
axis of the socket and is automatically plugged in if ungrabbed when
close to the socket.

Future additions to the connector system will include send/receive
objects as well as new PlugAttach mechanisms. The planned
SpherePlugAttach will snap the Plug to a sphere rather than a single axis.