Automate the arrangement of note cores in a chord
The arrangement of note cores in a chord is really not something that
needs to be configurable. The 'molecule' type structures that were
defined before look cool, but musically they were confusing. By
simplying stacking the note cores on top of one another with the lowest
pitch at the bottom and the highest pitch at the top, the NoteOrb
instantly conveys more useful information.

This change also introduces the possiblity of inverting
chords simply by flipping them over. The root note would stay the same
while the higher notes would move down an octave to be lower than the root note
and the lower notes would move up an octave to be higher than the root
note.