Improve MIDI note velocity detection
Instead of using VRTK controller events to detect when a note is
triggered, a custom script called ControllerNoteTrigger is used to
detect when a NoteOrb should turn on and off.

This script works by detecting when the trigger value stopped increasing
and emitting the NoteTriggered event. The note velocity is then
determined by how far the trigger is pressed in when this event occurs.
This makes it possible to play soft notes by lightly tapping the
trigger. Louder notes can be played by pressing it all the way in.

Currently, the velocity values are usable but the system isn't as
expressive as it could be due to some inconsistency with the velocity.

In order to find out how to improve these readings, some kind of
variable graphing is necessary. It's a fairly common problem to want to
see how a variable is changing over time when debugging, so it's
probably worth creating a custom editor window that can plot variables'
values.