Add Tilt Brush style cube menu
The main menu will be a Tilt Brush-like cube menu that has 4 panels. The
first panel will allow the user to grab and instantiate various modules
such as instruments, notes, effects, and controls.

The second menu will display information about the selected module. This
will be contextual based on the module, but for an instrument it will
include the option to change the instrument or load a preset. It should
also have a record arm toggle.

The third menu will show all the clips for the selected module and allow
the user to create physical clip triggers.

The design of the final menu isn't set, but it will likely be related to
scene navigation. If only 3 menus are needed then the menu shape will be
changed to a triangular prism.