Add PointerSelector, Player, SelectableModule
This commit lays the groundwork for the module selection system. The Player component will listen to events from newly created modules and the PointerSelector and update its selection accordingly. The Player component will pass its selection on to the SoundModuleMenu, which will be updated with the selected module's name. The menu will call the change instrument and load preset methods in the selected module when the corresponding buttons are clicked.