Add module selection system
Renamed Player to ModuleSelector, since the component only servers to
listen to events that result in a new module being selected and then
pass the selection on to the SoundModuleMenu.

The selection system brings up the issue of the various module types.
There will eventually be an AudioEffectModule, a MIDIEffectModule, and
an InstrumentModule. Currently the only module defined is SoundModule.

All of these modules will need to allow the user to browse for devices, load presets, select the module, and map module parameters to controls.

Therefore, there should be abstract base classes which define much of
the functionality currently laid out in SoundModuleController and
SoundModuleObject. These could just be called 'ModuleObject' and
'ModuleController'.

This means that a Module is a physical object that has a 1:1 correspondence with
a Bitwig track.

NoteOrbs and similar objects will no longer be considered Interface
Modules. Instead, they will called MIDI Instruments. There will also be
MIDI Control objects which do not output notes, but correspond to MIDI
CC messages that control instrument and effect parameters.