Remove most uses of ComposeVRManager Singleton
The ComposeVRManager was previously used to hand out references to certain global objects like menus and to store some shared variables. The global objects are now found using FindObjectOfType, while the shared variables are implemented using ScriptableObjects.

This is a more flexible implementation that won't result in a bloated
Manager class.