Remove blocker functionality from CordDispenser
There is no need for the CordDispenser to be blocked by other objects.

The blocker system gave no indication of what was blocking, so it was
confusing for the user. Now the player can at least see that the cord is
being dispensed into another object and move the dispenser.