Remove trigger logic from NoteOrb
A new component, TriggerableObject, must be attached to any object which can be played or triggered. Currently the NoteOrb is the only object with this component, but in the future ClipTriggers will also use this component.

The Baton is now responsible for looking for TriggerableObjects that it
is intersecting and sending them TriggerStart/TriggerEnd messages.