Change Batons to follow Actor pattern
The Actor gameObject should have the Actor component at the top of its
component list. This gameObject should be the highest in the Actor's
hierarchy.

The Baton component used to be on the head of the baton, but it has been
moved to the top level object. The head of the Baton is now an
ActorSubObject and has a SimpleTrigger that the Baton component
references to determine which Triggerables it should trigger.